+++ src/entrypoints/popup/routes/Themes.svelte
@@ -26,26 +26,6 @@
   let themes = themeSettings.defaultValue;
   let showModal = false;
 
-  // TODO
-  let logosAdaptive: LogoDetails[] = [
-    {
-      name: "Default",
-      url: "default",
-      id: "default",
-      disable: true,
-    },
-    {
-      name: "Catppuccin",
-      id: "catppuccin",
-      url: "https://raw.githubusercontent.com/catppuccin/catppuccin/main/assets/logos/exports/1544x1544_circle.png",
-    },
-    {
-      name: "Home",
-      id: "home",
-      url: "https://fonts.gstatic.com/s/i/materialiconsround/home/v16/24px.svg",
-    },
-  ];
-
   onMount(async () => {
     themes = await themeSettings.getValue();
     console.log("themes", themes);
@@ -69,25 +49,25 @@
 </script>
 
 <Modal bind:showModal>
-  <h2 slot="header" class="mb-10 text-xl">Choose an icon</h2>
+  <h2 slot="header" class="mb-4 text-xl">Choose an icon</h2>
 
   <div class="grid grid-cols-3 gap-4">
     {#each logos as logo (logo)}
       <button
         on:click={() => logoClicked(logo)}
         class:highlight={themes.logo.id === logo.id}
-        class="border border-ctp-pink p-2 flex flex-col items-center justify-between rounded-lg">
+        class="border border-ctp-accent p-2 flex flex-col items-center justify-between rounded-lg">
         <span>{logo.name}</span>
         {#if logo.disable !== true}
-          <img src={logo.url} alt="Logo" class="h-16 mt-2" />
+          {#if logo.adaptive}
+            <span class="logo-picker" style="--icon: url({logo.url})"></span>
+          {:else}
+            <img src={logo.url} alt="Logo" class="h-16 mt-2" />
+          {/if}
         {/if}
       </button>
     {/each}
   </div>
-  <br />
-  <!-- TODO -->
-  <h3 class="text-lg font-bold">Coming soon™</h3>
-  <p>Adaptive SVG logos that change based on your theme!</p>
 </Modal>
 
 <div id="card">
@@ -117,7 +97,7 @@
   <button
     title="Choose icon"
     id="choose-icon"
-    class="flex items-center mx-2 small hover:bg-ctp-pink hover:text-ctp-crust"
+    class="flex items-center mx-2 small hover:bg-ctp-accent hover:text-ctp-crust"
     on:click={() => (showModal = true)}>
     <Layers3 />
     <span class="ml-3">Choose an icon</span>