+++ src/modules/plugins/modernIcons/index.ts
@@ -29,6 +29,9 @@ export default async function defineWxtPlugin() {
       ["icon-group", "group"],
       ["icon-info", "info"],
       ["icon-resource-booking", "book_online"],
+      ["icon-files", "description"],
+      ["icon-schoolbox", "language"],
+      ["icon-user", "person"],
     ];
 
     function insertIcon(className: string, iconName: string) {